fix OpenSSL function signatures

+++ openssl/rand.go
@@ -15,7 +15,7 @@ type randReader int
 func (randReader) Read(b []byte) (int, error) {
 	// Note: RAND_bytes should never fail; the return value exists only for historical reasons.
 	// We check it even so.
-	if len(b) > 0 && C.go_openssl_RAND_bytes((*C.uint8_t)(unsafe.Pointer(&b[0])), C.size_t(len(b))) == 0 {
+	if len(b) > 0 && C.go_openssl_RAND_bytes((*C.uint8_t)(unsafe.Pointer(&b[0])), C.int(len(b))) == 0 {
 		return 0, fail("RAND_bytes")
 	}
 	return len(b), nil
